Procesa API y crea estado global thunk
