+++ src/Store/index.js
@@ -1,0 +1,8 @@
+import { configureStore } from "@reduxjs/toolkit";
+import partidos from './slices/fixtures.slice'
+
+export default configureStore({
+    reducer: {
+        partidos,
+    }
+})
