+++ src/components/HeaderNav/styles/HeaderNav.css
@@ -1,0 +1,7 @@
+.headernav__container {
+    display: flex;
+    flex-direction: column;
+    gap: 1em;
+    flex-wrap: nowrap;
+    margin: 0.5em;
+}
