+++ src/utils/apiConfig.js
@@ -1,0 +1,2 @@
+const API_BASE_URL = "http://64.23.243.233:6969/v1/fixtures/"
+export default API_BASE_URL;
